detection: add initial form classification
Add initial support for form classification of targets as described in
the AUVSI conops.
This classifier uses a Convolutional Neural Network to classify a 32x32
RGB image as one of fourteen ouput classes (circle, half circle, ...).
Training this model requires a large dataset of images which can be
generated with auvsi-targets[1].

[1] https://github.com/amitibo/auvsi-targets

Signed-off-by: Liam Beguin <liambeguin@gmail.com>